Kafka/consumer/requirements.txt
kafka-python
python-dotenv
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly